refactor: card 패키지 삭제 후 game 패키지로 이동
