+++ src/main/java/strategy/DeckShuffleStrategy.java
@@ -1,7 +1,7 @@
 package strategy;
 
-import card.Card;
-import card.Cards;
+import game.Card;
+import game.Cards;
 import constant.TrumpEmblem;
 import constant.TrumpNumber;
